chore!: use u64 for NoiseLevel
Change from usize to u64 for MaxNoiseLevel and NoiseLevel

This is an API break as `new` and `get` handle/returns u64
instead of usize

This is also a potential serialization break depending on the
serializer used (bincode should be fine as it serializes usize as u64)

+++ tfhe/src/c_api/shortint/parameters.rs
@@ -62,7 +62,7 @@ impl TryFrom<ShortintPBSParameters> for crate::shortint::ClassicPBSParameters {
                 c_params.modulus_power_of_2_exponent,
             )?,
             max_noise_level: crate::shortint::parameters::MaxNoiseLevel::new(
-                c_params.max_noise_level,
+                c_params.max_noise_level as u64,
             ),
             log2_p_fail: c_params.log2_p_fail,
             encryption_key_choice: c_params.encryption_key_choice.into(),
@@ -113,7 +113,7 @@ impl ShortintPBSParameters {
             ks_level: rust_params.ks_level.0,
             message_modulus: rust_params.message_modulus.0,
             carry_modulus: rust_params.carry_modulus.0,
-            max_noise_level: rust_params.max_noise_level.get(),
+            max_noise_level: rust_params.max_noise_level.get() as usize,
             log2_p_fail: rust_params.log2_p_fail,
             modulus_power_of_2_exponent: convert_modulus(rust_params.ciphertext_modulus),
             encryption_key_choice: ShortintEncryptionKeyChoice::convert(
